+++ src/main/java/com/example/onlineshop/entity/order/Order.java
@@ -19,10 +19,10 @@ import java.util.List;
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer orderId;
+    private Long orderId;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 //    @ManyToOne
 //    @JoinColumn(name = "user_id", insertable = false, updatable = false)
 //    private User user;//раскомментить когда добавят юзера
@@ -40,19 +40,19 @@ public class Order {
     @Column(name = "comment")
     private String comment;
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -88,7 +88,7 @@ public class Order {
         this.condition = condition;
     }
 
-    public Order(Integer orderId) {
+    public Order(Long orderId) {
         this.orderId = orderId;
     }
